requirements.txt
pytz==2021.3
influxdb==5.3.1
slackclient==2.9.3
slack_bolt==1.11.0

